Ensure that JIDs in roster item stanzas are stripped
This is a touchy issue -- a literal read of RFC6121 doesn't 
preclude adding roster items with a resource component, although
numerous discussions suggest that this is due more to pedantism
than any typical usefulness. This patch ensures that JIDs are
stripped before going over the wire, so that we only
ever talk in bare JIDs to the server.

Note that because this is done in the stanza and not in the 
roster itself, you're free to store items in your local 
Roster however you'd like. All this changes is what's on the wire.