delegate to_xml (with options) to the node
Instances of Blather::StanzaError and Blather::StreamError had an
implementation of `to_xml` that does not take an argument.  A problem
with the previous implementation is that the following statement will
throw an exception:

    client.write Blather::StanzaError.new(stanza, 'registration-required', :cancel)

The reason is that client.write calls stream.send, which in turn calls
to_xml with an argument on the stanza:

    stanza.to_xml(:save_with => Nokogiri::XML::Node::SaveOptions::AS_XML)

The difference between `to_xml` with the formatting option and `to_s` is
whitespace.  The string produced with `to_xml(:save_with => AS_XML)`
strips off any whitespace between tags whereas `to_s` keeps the
whitespace between tags intact.