Fix unsafe threaded use of @tmp_handlers in Blather::Client
Plain hashes are not thread-safe, especially under JRuby. This caused
us a fair bit of misery with Adhearsion in production. Acknowledgement
of commands issued to FreeSWITCH would occasionally go unnoticed,
resulting in calls freezing and timing out after one minute.

It does not appear to be necessary to apply this to @handlers, which
is also a plain hash.